fix: address round 4 PR review suggestions
