+++ src/__tests__/components/pwa/usePwaInstall.test.ts
@@ -24,6 +24,10 @@ describe('usePwaInstall', () => {
         });
     });
 
+    afterEach(() => {
+        jest.useRealTimers();
+    });
+
     beforeEach(() => {
         Object.defineProperty(navigator, 'userAgent', {
             value: 'Mozilla/5.0 (iPhone; CPU iPhone OS 17_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/17.0 Mobile/15E148 Safari/604.1',
@@ -103,6 +107,5 @@ describe('usePwaInstall', () => {
             jest.advanceTimersByTime(30_000);
         });
         expect(result.current.showBanner).toBe(true);
-        jest.useRealTimers();
     });
 });
