+++ src/__tests__/lib/pwa/detectPwaEnvironment.test.ts
